more item datamodel progress
