+++ static/template/actor/tab/vehicle/main.hbs
@@ -110,9 +110,9 @@
               <img class="item-property item-image" src="{{item.img}}"></img>
               <div class="item-property item-name">{{item.name}}</div>
 
-              <div class="item-property">{{item.Damage}}</div>
-              <div class="item-property small">{{item.ED}}</div>
-              <div class="item-property small">{{item.AP}}</div>
+              <div class="item-property">{{itemsystem.damage.formatted}}</div>
+              <div class="item-property small">{{item.system.damage.ED}}</div>
+              <div class="item-property small">{{item.system.damage.AP}}</div>
               <div class="item-property">{{item.Range}}</div>
               <div class="item-property flex">
                 {{#if item.isRanged}}
@@ -127,9 +127,9 @@
                 {{/if}}
               </div>
           </div>
-          {{#if item.traits.length}}
+          {{#if item.system.traits.list.length}}
           <div class="item-content-extra">
-            <span class="item-traits"><b>Traits:</b> {{{array item.Traits "item-trait"}}}</span>
+            <span class="item-traits"><b>Traits:</b> {{{array item.system.traits.formatted "item-trait"}}}</span>
           </div>
           {{/if}}
         </div>
