+++ static/template/item/ability.hbs
@@ -203,7 +203,7 @@
           <div class="item-form-group stacked">
             <label>{{localize "WEAPON.TRAITS"}}<a class="item-traits"><i class="fas fa-cog"></i></a></label>            
             <div class="item-form-inputs">
-              <input disabled type="text" value="{{item.Traits}}" />
+              <input disabled type="text" value="{{item.system.traits.formatted}}" />
             </div>
           </div>
         </div>
