+++ static/template/item/armour.hbs
@@ -80,7 +80,7 @@
           <div class="item-form-group stacked">
             <label>{{localize "ARMOUR.TRAITS"}}<a class="item-traits"><i class="fas fa-cog"></i></a></label>            
             <div class="item-form-inputs">
-              <input disabled type="text" value="{{item.Traits}}" />
+              <input disabled type="text" value="{{item.system.traits.formatted}}" />
             </div>
           </div>
         </div>
