+++ static/template/item/weapon.hbs
@@ -107,15 +107,15 @@
           <div class="item-form-group fixed">
             <label>{{localize "WEAPON.ED"}}</label>
             <div class="item-form-inputs">
-              <input name="system.ed.base" class="base" type="number" value="{{system.ed.base}}" data-dtype="Number" />
-              <select name="system.ed.rank" class="rank">
-                {{#select system.ed.rank}}
+              <input name="system.damage.ed.base" class="base" type="number" value="{{system.damage.ed.base}}" data-dtype="Number" />
+              <select name="system.damage.ed.rank" class="rank">
+                {{#select system.damage.ed.rank}}
                 <option value="none">{{localize "RANK.NONE"}}</option>
                 <option value="single">{{localize "RANK.SINGLE"}}</option>
                 <option value="double">{{localize "RANK.DOUBLE"}}</option>
                 {{/select}}
               </select>
-              <input name="system.ed.bonus" class="modifier" type="number" value="{{system.ed.bonus}}" data-dtype="Number" />
+              <input name="system.damage.ed.bonus" class="modifier" type="number" value="{{system.damage.ed.bonus}}" data-dtype="Number" />
             </div>
           </div>
         </div>
@@ -123,15 +123,15 @@
           <div class="item-form-group fixed">
             <label>{{localize "WEAPON.AP"}}</label>
             <div class="item-form-inputs">
-              <input name="system.ap.base" class="base" type="number" value="{{system.ap.base}}" data-dtype="Number" />
-              <select name="system.ap.rank" class="rank">
-                {{#select system.ap.rank}}
+              <input name="system.damage.ap.base" class="base" type="number" value="{{system.damage.ap.base}}" data-dtype="Number" />
+              <select name="system.damage.ap.rank" class="rank">
+                {{#select system.damage.ap.rank}}
                 <option value="none">{{localize "RANK.NONE"}}</option>
                 <option value="single">{{localize "RANK.SINGLE"}}</option>
                 <option value="double">{{localize "RANK.DOUBLE"}}</option>
                 {{/select}}
               </select>
-              <input name="system.ap.bonus" class="modifier" type="number" value="{{system.ap.bonus}}" data-dtype="Number" />
+              <input name="system.damage.ap.bonus" class="modifier" type="number" value="{{system.damage.ap.bonus}}" data-dtype="Number" />
             </div>
           </div>
         </div>
@@ -140,7 +140,7 @@
           <div class="item-form-group stacked">
             <label>{{localize "WEAPON.TRAITS"}}<a class="item-traits"><i class="fas fa-cog"></i></a></label>            
             <div class="item-form-inputs">
-              <input disabled type="text" value="{{item.Traits}}" />
+              <input disabled type="text" value="{{item.system.traits.formatted}}" />
             </div>
           </div>
         </div>
